// src/models/IEntidad.ts
export  class IEntidad
{
    Cordenadas: {Latitud: number, Longitud: number}
    Correo: string
    Descripcion: string
    Direccion: string
    Lema: string
    Negocio: string
    Nombre: string
    PaginaWeb: string
    ServicioADomicilio: boolean
    Telefono: string
    UrlImagen: string    
}